requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
importlib-metadata==4.8.2
Markdown==3.3.6
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
zipp==3.6.0
